Move file cleanup to ZTestRunnerJVM
